simplify and add tests
